Remove old examples
